requirements.txt
streamlit
pandas
plotly
google-generativeai
numpy
scikit-learn